Renamed source files where classes are defined to match class names.
